requirements.txt
beautifulsoup4
discord.py
pydantic
databases
aiosqlite
